📛rename:  상태관리 훅 폴더 이동 - hooks 에서 store로 이동
