+++ src/components/modals/SignUpModal.tsx
@@ -5,7 +5,7 @@ import { RiKakaoTalkFill } from '@react-icons/all-files/ri/RiKakaoTalkFill';
 import Image from 'next/image';
 import Link from 'next/link';
 
-import useSignUpModal from '@/hooks/useSignUpModal';
+import useSignUpModal from '@/store/useSignUpModal';
 
 import Modals from './Modals';
