+++ src/store/useSearchErrorModal.ts
